feat(permissions): introduce `applicable_area` on access levels
Users access the application from three distinct situations: Either a
public portal (for publications), as applicants, or internal.

Most access levels should only apply in one of these situations or areas
of the application. For example, someone may submit a building permit dossier
while also being employed at the responsible municipality. This person
should then have "applicant" permissions in the portal, but should not
see the dossier internally while it's not submitted yet.

Similarly, when the same user is viewing the publications, their dossier
must not show up there before it is actually being published.

The `applicable_area` defines for each access level where it should
be considered for deciding access to a dossier (bot visibility and the
associated permissions). If the user is working in the internal area of
the application, any "PUBLIC" or "APPLICANT" ACLs will not be checked,
and vice versa.