feat(permissions): add mode switcher utilities
During the transition, where the permissions module is integrated all over the
place, we need a way to switch between old mode and new mode permissions.

The switcher allows adding methods that help during the transition, by allowing us
to run in various modes:

* In "FULL" mode, the permissions switcher will *only* use the permissions
  module, the "traditional" permissions are not checked at all. This will
  be the default once the transition is complete and we are sure everything
  is correct.

* In "CHECKING" mode, both backends are checked, and any difference is treated as
  an error. This should be used during development of the transition.

* In "LOGGING" mode, the same thing happens as in "CHECKING", but any difference
  is only logged as warning, but not treated as an error (so it won't trigger
  HTTP/500 errors, for example).

* In "OFF" mode, only the old permissions code is used. This can be used for
  production while parts of the transition are already merged, but not
  considered stable enough to activate. *This is the default!*

There is a fifth mode as well, which will be handy exactly once - when we
run through the code and clean everything up once we're sure we've migrated
properly. This won't be used until then.