feat(permissions): add require_any() and require_all() methods
These do exactly what you think: If the user has any of the
permissions you pass to `require_any()`, nothing happens, otherwise a
`PermissionError` is raised. The same happens with `require_all()`, but
all permissions will be required, not just one.

For convenience, `has_any()` and `has_all()` methods are available as well.