feat(permissions): begin migration tooling
The permissions migration tool is creating / revoking instance ACLs
such that they reflect the data which was previously used to define access
(Such as being the active service, an applicant, etc) for an instance.

This tool is a work in progress, and will be extended as more and more
functionality is moved from "traditional" access control to the permission
module.