feat(kt-bern/permissions): introduce "applicant" access level
This access level is used to define the permissions that an applicant has
on the dossier. It will certainly be extended in the future, this is WIP,
at least during this merge request.

Also implement the event handlers for dealing with adding/removing
applicants: If an applicant is added or removed, grant or revoke the
corresponding ACLs.

The applicant event methods are in a new "common" module, so other
cantons can easily reuse them. Goal is to have as much event handler
code in the "common" module as possible.