feat(applicants): trigger permissions events
When applicants are added or removed, we trigger events so that
application code can grant/revoke corresponding instance ACLs.

Applicants are special: Users can invite people by email, without the
invitee having an account in the system. When the user is logging in the
first time, any "email" applicant is turned into a "real" applicant.

The "applicant_added" event is triggered whenever an applicant is added,
not only manually, but also on instance creation, and when one without
an user account is turned into a "real" applicant.

This way, there should be no "lost" applicants.