chore(communications): remove obsolete create attachment API
This API was implemented because of a misunderstanding of the
specification. Later on, the creation of communications attachments was
implemented as multipart on the creation of the message.

However, the old API was never removed and still exists with a lot of
unnecessary complexity. This commit removes the old obsolete API parts
so we don't have to maintain unused code anymore.