refactor(dossier-import): make test setup for config more reusable
Unify testing params that are shared among projects and make
differences more visible. This might not be the best way to go
about it. It allows however to keep a rather good record of
changes and keep track of respective snapshots that tend to be
overwhelming and changes might be overseen