fix(permissions): return old data in logging mode
The permissions mode switcher would return the "NEW" value in "LOGGING" mode.
This is only useful during development of the permissions module, but not
for people working on non-permissions-module code.

The LOGGING mode now returns the old data (but still logs the
difference). A new DEV mode is introduced that has the old behaviour:
It logs any difference, but returns the new data. This is to be used
for permission module development.