fix(tests): do not crash in public url test (flaky test)
The ECH0211 view uses the instance queryset mixin, which will trigger an
invalid test situation in the public urls test. Depending on test order, the
condition may not be triggered.

The fix disables the ECH0211 interface and reloads the URLs, so the
public URL test doesn't check any URL that's not working that way