fix(be/permissions): do not crash when creating instances
When creating an instance, the BE permissions trigger would create an
instance ACL for the creator (with "applicant" permissions).

This in turn would try to send notifications to the user referenced by the
new ACL, which might include the Geometers (but we can't know yet). To
find this out, the code would try to lookup the responsible municipality, but
because it's a new dossier, this information would not yet exist.

The fix is two-fold:

A) explicitly disable notification on added applicants
B) catch exceptions during the notification and log a warning instead of
   crashing