feat(permissions/municipalities): support involved authority, change of lead
There are a few things that need to happen for this:

* Introduce access level (and configuration) for involved authority
* Implement event trigger when lead authority changes
* Change the migration tooling to support the new access level as well

When changing the lead authority, grant the "old" authority the
`involved-authority` access level and revoke the `lead-authority` ACL at
the same time, and grant the new lead  `lead-authority` access.

The same behaviour needs to be mapped in the migration script as well,
now mapping inactive instance services to `involved-authority` and active ones
to `lead-authority`.