fix(instances): make sure queryset is distinct
We use the `InstanceQuerysetMixin` in the instance serializer but
override the default `get_base_queryset` method of the mixin. In that
method of the mixin we use `.distinct()` in order to avoid duplicates
caused by the permissions module.

To avoid any errors, we no use `.distinct()` on the overwritten method
in the serializer as well.