feat(permissions): only show available access levels in create-acl modal
When creating ACLs, we should only see the access levels that are actually
available for creating ACLs. The permissions module already uses this to
decide whether to allow or disallow creating of ACLs:

* `permissions-grant-any`: With this permission, you are allowed to grant
  ACLs using all access levels
* `permissions-grant-FOO`: With this permission, you are allowed to grant
  ACLs using the `FOO` access level (the slug must match)

For this, introduce a filter in the backend (assignable_in_instance)
and use that filter in the frontend when listing the available access levels
in the "create ACL" modal.