fix(permissions/filters): make the assignable_in_instance filter behave correctly
The filter should behave correctly depending on the permission module's configured
operating mode: If it is in FULL mode, you are 100% required to have the required
permissions via your access level/ACL.

If not in FULL mode, the municipality users see everything, and the rest don't see
anything.

Also move the `test_assignable_filter` to where the other access level tests are.