feat(dev-env): instrument coverage and find tests for given file/line
Pytest-Coverage has a feature that allows recording dynamic contexts,
which it will name after the current test function.

This data in turn can then be examined to figure out which test is actually
covering which line. Very useful for two things:

* First, figure out which tests to run to test a given piece of code
* Second (and this is kinda not working / to-be-discussed):
  Figure out if a line of code is tested purposefully or accidentally.
  The idea being that each line of code needs to be tested by a test
  case local to the given module. For example we do not (want to) accept
  coverage in the permissions module if the code is only run by tests in
  some other module that likely does not care about it

While at it, discovered some issues with our module structure: Not all
module directories have an __init__.py file, which causes the identification
of some tests to be represented incorrectly (because Python would not
consider the directory as a proper module)