feat(permissions): introduce role-type acls / access levels
Some groups of users cannot be identified by service, but only by role.
The "support" type in BE is one of them, or KOOR services in UR.

Therefore, we now add a "role" type ACL, and extend the access levels to allow
role type ACLs as well.

While we're at it, refactor the ACL validation a bit, so it becomes less
repetitive.