dev(django-q): add global sync option to django-q config
and refactor dossier-import documentation

Running django-q async tasks in sync mode for testing does not
work on a per task basis in tests but requires the global setting
which needs to be set at startup time.

E. g. overriding the setting on a per test basis in test setups
has no effect.

NOTE: differing setup of the env var `DJANGO_Q_ENABLE_SYNC` is due to
the fact that some are setting environment as mappings some as list.
mappings need a value. lists can merely mention the variable name
to pick it up from the `.env`.