feat(dev): order dump filter groups by condition complexity
Currently, we fill the dump groups in the order that they are
configured. This works fine in basic use-cases but it will fail as soon
as we configure more complex conditions where a model could go into
multiple groups.

A simple use-case for this would be that we have group
for a v2 of the main form but also a v2 of the decision form. Right now,
the questions for the v2 decision form would be put into the v2 main
form dump as this group is defined before the new decision v2 group.

With this change the dump command will determine how complex the
condition for the dump group is and will prioritize the more complex
ones.