chore(construction-monitoring): refactor dynamic tasks
We always need to return a list of strings rather than a queryset of
task pks. Otherwise Caluma will raise an error if this dynamic task is
followed by another task. Caluma will try to merge and flatten a
queryset of task pks with a list of strings which will fail.