fix(be): remove "user" user from Guest and Portal groups
The behaviour of the internal area becomes awkward¹ if we select the
Guest or Portal group. But also there is no reason to select those
groups in the first place, because they are only used for
lower-level/system purposes.

To prevent curious users/developers from accidentally selecting the
group, we remove them from the default set of assigned groups for the
"user" user.

¹ Redirection to login screen and no way to log in back to the internal
  area.