fix(Makefile): replace/improve echo usage and escape sequences
echo's handling of -e, -n and backslashes in strings is inconsistent in
bash, dash, zsh and the version provided by GNU coreutils, which is why
printf is generally the recommended option for scripts or any other
non-trivial use of string printing.

Also \e is not a standard escape sequence and thus not recognised by all
implementations of echo/printf, so we use the numeric/octal notation
(\033) instead.