feat: test advisor utility
This is a new tool that should help keeping track of changes across
a release specifically for testing.

The idea is to run this across the Git commits since the last release,
and it will figure out which modules changed, which cantons are affected,
and therefore what will need to be tested.

There is a configuration file that defines the relevant files for each
module, and it is possible for each module to denote other modules that
may be affected by it.

Note: This is still rather crude at the moment, and the configuration is
nowhere near complete. This is a first draft to see whether such a tool
might be useful if brought to completion.