feat(dossier-import): import responsible user
The responsible user is set in the "responsibles" module, as follows:
If the user is found, it is written as the currently responsible user,
using the instance's instance service. An assumption is made that the
user belongs to that service, no validation is performed.

If the value in the Excel column is "<LOESCHEN>" or `<LÖSCHEN>`, then
any existing value is removed. If the value is empty, no change takes place.