feat(permissions): add separate view to get permission module meta info
Currently, the frontend fetches the instance permission endpoint with a
page limit of 1 to get the permission mode. For some reason, this
endpoint sometimes returns a 403 error. As the frontend really only
needs to know the permission mode without any information about an
instance, we add a separate endpoint which only returns that
information.