fix(gr): no documents-write permission while module is not enabled
In Kt. GR, we currently only use the permission module for the "read"
access type. Since lots of the permission logic is already in place, we
already generate instance ACLs based on events, though. This currently
breaks the logic in the alexandria permission layer, which assumes that
only "read" permissions are present in the database. This fix
temporarily removes any "documents-write" permission for non-"read"
access types, which fixes the Alexandria permission layer for GR until
the permission module can be fully enabled.