fix(communications): handle invalid entity value
If we get a null (or forged) entity from the frontend, we might run into
None values. If unchecked, these would lead to unhandled exceptions and
500 errors in the backend.

Also remove the "fixme" comment about hardcoding the involved entities:
it is better to error out in case we get a request to create a topic to
inacceptable entities (rather than guessing and assuming and opening a
topic to a "wrong" LB).