fix(dossier-import): remove uncompress flag for mimetype detection
The `uncompress` flag was active and lead to errors when importing, for
example, docx files. In my understanding `uncompress=True` should only
be used when giving a file path so the Magic library can uncompress and
read the file itself. But since we are reading the file and passing the
content of the file to the Magic library the `uncompress` led to errors
as it tried to uncompress the docx content further. Therefore
deactivating `uncompress`.