fix(permissions/be): allow municipalities only to grant geometer
BE currently only uses the permission module for granting access to
Geometers. To avoid abuse, we should make the `assignable_in_instance`
filter such that only the assignable access levels are listed.

Additionally, the `enforce_change_permission()` method in the view now
checks the same filter to see if one can actually assign a an access
level on a given instance.