fix(cases): ensure only top-level cases are fetched in case list
In certain (unknown) circumstances, multiple inquiry child-cases
are generated, which results in one of the child-cases not being
attached to the parent work-item. The existing case list filter
only checks whether the case isn't attached to a parent work-item.
This resulted in broken case lists with inquiry child cases that
weren't handled properly.