fix(token-exchange): improve frontend handling of multiple sessions
This commit fixes a number of issues when handling OIDC and token
exchange sessions in the portal. This happens as users can still use the
portal with regular OIDC for paper instances. The change makes sure that
existing sessions are properly invalidated if a new session of a
different type (token exchange / OIDC) is started.