fix(documents/performance): greatly improve thumbnailing speed
Our thumbnailing engine uses imagemagick behind the scene, which
by default looks at the whole document. It has a performance issue
that is dependent on the physical size of the document. SORL
therefore only looks at the first page for calculating thumbnails,
but to detect the page orientation, this tiny suffix is missing.

Therefore, documents with a huge amount of pages will take a significant
time to thumbnail, all due to the fact that it looks at all the pages
instead of just the first one.