fix(sz): correctly send notification to tax admin when needed
settings.TAX_ADMINISTRATION is used to send the email when completing an
instance (assuming the checkbox is selected). The ID is incorrect, so
fix that.

This also means we shouldn't send out the notification unconditionally
based on config setting (NOTIFICATIONS/NOTIFICATION_SETTINGS), but leave
it up to camac.notification.NotificationTemplateSendmailSerializer.
_get_recipients_tax_administration.