fix(sz): use correct service IDs for tax administration
This partially reverts 9a6b8e3b9e2dce1ce0cd4ff0914661c8da751fb7, as the
"fix" there was incorrect: 585 is the *group* ID for
Liegenschaftenschätzung, which is not what we need here.

Also the tax administration is a different organisation altogether, so
treating the two as the same is also wrong.

This did not end up in emails being sent to the wrong recipients, but
rather no emails being sent at all (as there are no organisations with
that ID).