fix(sz): use correct service IDs for tax administration
This partially reverts 41d0677ed7a0aab6711990db431ac70df2ff5294, as the
"fix" there was incorrect: 585 is the *group* ID for
Liegenschaftenschätzung, which is not what we need here.

Also the tax administration is a different organisation altogether, so
treating the two as the same is also wrong.

This did not end up in emails being sent to the wrong recipients, but
rather no emails being sent at all (as there are no organisations with
that ID).