fix(work-items): fix links for completed work items
Currently, completed work items do not have any links to the detail
view. This is a regression as this worked at some point and should
absolutely work in order to be able to see information about who
completed the work item at what point.

For manual work items this is even more crucial as the person that
completes the work item can add additional information about how they
completed it. This information can't be seen anywhere right now.

Related to kt-so/ebauso-coordination#341