fix(performance): cache landowners data source
Cache the landowners data source to ensure that during
validation of the document (for example during PDF generation),
the landowners aren't retrieved for every single table row
of the parcel table (which leads to considerable computation times).
The short cache timeout ensures that even when adding new landowners,
the up-to-date data will be retrieved by the time the parcel table
is reached.
In addition, re-enable the visibility checks for the landowners.