feat(admin): simplify creation of new services in django admin
Currently the creation of a new service (which happens a lot for new
cantons) is really painful as you have to create the service, then
create all groups and make sure to always use the correct naming.

This can be simplified as the required groups are static per service
group. This MR automates the creation of those groups and also syncs the
correct generated group names to also simplify service name changes.

Related to kt-so/ebauso-coordination#350