refactor(alexandria): use group from top level request object
Using the group and user information from `request.caluma_info` doesn't
make sense as alexandria has nothing to do with caluma and doesn't need
any data that is not already available on the regular django request.
This MR simplifies the usage of the request group alot which will make
using alexandria permissions and visibilities in other parts of the
application alot simpler.