fix(ech0211): don't check document visibility if queryset is empty
This is only a temporary fix that will be reverted when the actual
problem is solved. Currently the documents passed in eCH-0211 events are
always `Attachment` models even if the canton uses Alexandria. This must
be fixed. In order to provide a temporary fix we just ignore any
visibility rules if the queryset is empty which it will always be as a
canton using Alexandria will never have any legacy attachments.