fix(sz): match instance personalities only by unique tuple
The (instance, applicant_user) tuple must be unique, so get_or_create
with additional properties might lead to conflicts.

One way to solve this would have been to pass the user and email
properties as `defaults` dictionary: this would have properly matched
the entry and left it unchanged if it already existed.

Another way (which we choose here) is to do the same but with
update_or_create: we now instead *update* the entry with the new
information if it is matched.

Rationale: The email information is outdated anyway, both the current
applicant_user and the personality have been declared with the new email
address, so there is no point in keeping around outdated information.