fix(dossier-import): always parse row IDs as strings
Excel will interpret a value of "133" as an integer but a value of
"2011-133" as string. If the ID is interpreted as an integer, our
validation will fail saying that the corresponding folder doesn't exist.
Since those IDs will be saved as tags / keywords in a text field, it
makes sense to always cast them to a string in order to avoid
mismatches.