refactor(documents): permissions do the filtering now
The permission objects in the documents module do the filtering now,
so we don't have to split out the permission classes by "internal" and
"full section" anymore to determine what's readable.

This is not yet a full refactor of what *could* be done, just simplifying
stuff so we can later on integrate permission-module access levels.

The change in the attachment deletion test is due to a bug in the previous
implementation: It should not only be forbidden to delete that attachment,
but it should not be visible at all. So the change from HTTP/403 to 404 is
fixing a bug here