feat(documents): implement access level based visibility
This is currently implemented in an "OR" fashion, and does not disable
the visibility defined by the other (role-based) permissions.

We also do not override the `@permission_aware` specialisations such
as `_for_geometer` and `_for_public`: These still function just as before.

The now-higher query counts in the `test_attachment_list` test are due
to the fact that we're loading the user's relevant access levels from
the DB. While it's an additional query, it should not have a big impact,
as it's rather simple.