test(fixtures): generate code for module settings fixtures
This removes a whole lot of black magic in favor of explicity. This will
help us find where those previously "magic" module settings fixtures are
coming from. Also, the generate command uses the actual configuration to
determine whether a module is enabled in a canton so we don't have to
update a list of enabled cantons per module.