fix(permissions): fix displaying of entity email in details modal
This used to work when we used the non-public entities for service and
user as relationship. However, we switched to the public entities so we
can use the models in the portal as well. Those do not contain the
entity email which were then never shown in the detail modal.