fix: strip whitespaces from personalities' emails
If a user mistakenly adds a leading/trailing whitespace to an otherwise
correct email address, this causes the matching for a corresponding user
account to fail when creating the Applicant object.

This results in invitees unable to access an instance with their user
account despite having received a notification (note that the
notification still reaches its destination because send_mail also does
`strip()` on all recipient email addresses prior to sending).