fix(sz): do not update existing applicants when existing
As the `user` property may be set to someone else than the instance user
(due to being directly invited as involved participant), we should not
overwrite this property upon submission.

This leads to the potentially outdated `email` property not being
updated either, but this property is not very relevant (at least for
registered users).