fix(sz): look up and create involved applicants manually
Depending on whether a corresponding user exists for a given email, the
lookup for "do we need to create a new applicant?" is different:o

 * If the user exists, we're interested in the (instance, invitee)
   tuple, which should be unique.

 * If the user does not exist, we're interested in the (instance, email)
   typle, which should also be unique.

The previous approaches (with get_or_create/update_or_create, with the
'lookup tuple' always the same regardless of the situation) has been too
rigid.

The new approach first manually looks up the user, and depending on its
existence, it looks up the applicant using one or the other tuple.

And it then creates the applicant if missing.