fix(sanctions): do not set createdByService on template creation
There appears to be a timing issue that causes peekRecord() to return
null on staging, which in turn causes the backend to reject the
parameter (it accepts any service associated with the request user, but
not null).

But really, there is no need to pass this from the frontend, as we're
setting it in the backend anyway, so we remove it altogether.