feat(ur): enable sanctions
This adds an instance filter for filtering for instances with pending
sanctions or even pending sanctions assigned to a specific service.

This also introduces a sanctions settings module to limit the services
to which sanctions can be assigned.

It also unhides the sanction resources in the camac-ng internal area,
enables the sanctions data source for use in the construction
monitoring, and updates the construction monitoring form to use new
dynamic multiple options questions in place of the old sanction
confirmation question.