feat(utils): add a retry utility
This can be used for operations that might fail, but usually succeed at the
second try. The given callable will be run multiple times if an exception
is thrown, but if it succeeds, it's return value is returned.

This is used for example in GEVER, where if we create objects too fast
in their API, they're not *actually* created just yet, and we'll need
to wait a tiny bit and try again