feat(be-gever): initial infrastructure code for gever interface
Implementation for connecting to their REST API, including authentication.

Add a data model representation of the API objects so they become easier
to work with (and hopefully allow some simple validation as well).

Also introduce DB models for some configuration that's too complex to
track using settings, as it's differing a lot across the CMI environments.

In this commit, we mainly concentrate on talking to the CMI API, and leave the
integration/synchronisation bit for later. One exception is the addition of
a Caluma data source to fetch the "Erledigungsart" from GEVER, which is needed
to build up the corresponding Caluma form.