feat(ag): add backend for responsible user rules
Related to DIB-385

This adds the backend for the new module to create responsible user
rules. Those rules are being evaluated per service when a service is
newly involved in an instance. Currently, this is being triggered when a
new InstanceACL is created as this feature is only used in AG where the
permission module is fully enabled.