feat(ag): add frontend for responsible user rules
Related to DIB-385

This adds the frontend for the new module to create responsible user
rules. The module is intentionally scoped inside a "rulesets" module
that will be extended with another submodule to define custom deadline
rules.

This commit contains a sortable list view and views to edit and create
responsible user rules.