feat(be/keycloak): apply CI to all keycloak pages
The canton requested to apply the default CI style to some keycloak
pages and further wishes for a custom layout of the login-update-profile
page. To apply the general look and feel we moved the template-be.ftl to
template.ftl which serves now as a template for all keycloak pages.

Further we introduced UIkit to apply custom styles to some pages and to
match the general look and feel of the eBau application.