feat(sz): show Geometer field in Projektgenehmigungsgesuch
As the question must only be visible for municipality and district
streets (and not cantonal streets), we additionally need to check the
`fieldType` meta property. As the `meta` field is a JSON value, we also
add a JEXL transform for parsing JSON expressions.

This does not affect existing Geometer fields (the active-expression
only hides the field conditionally on Projektgenehmigungsgesuch forms,
which is new), so we can keep using the existing geometer-v4 question.
We do need a new version for the form itself, however (as we're adding a
new question there).