fix(sz): remove gesuchsunterlagen-v2 subform
It was originally created for adding a new info question [1], but that
is unnecessary: gesuchsunterlagen is a bit of a special form because it
only shows info messages related to questions/answers in a previous step
of the form.

So there is no need to create a new version because new info questions
won't show up on older forms (where the related question doesn't exist).

[1] 948024c31be73a3bf97e0be96def713db2dd53d8