chore(sz): split up case-based filtering for inquiries
Instead of having one method that tries to match both root cases and
distribution cases, we now provide two distinct methods, and the calling
code must (based on the context it knows about) invoke the right filter
method.

This makes the individual methods simpler, and avoiding the `Q(…) |
Q(…)` also avoids the `OR` clause in the generated SQL statement that
causes the query to run significantly slower (for some yet unknown
reason).