feat: save instance copy source information on instance
The semantics is that we are interested in knowing whether an instance
was copied from another instance, so we reflect that on the instance
itself instead of relying on some Caluma internals.

This reduces the number of tables to look up (and the number of joins),
which is especially noticeable in the permissions layer.