feat(so): hide cantonal inquiries for municipality
All inquiries which the current service is neither addressed
or controlling and the current service is not cantonal or bab
and the inquiry is not from a sibling or parent service and
the service in controlling_groups is cantonal are now hidden.

We also split up the additional_inquiries_filter into a separate
function to reduce overall function complexity.