feat(sz): remove location criteria for geometer
For the currently active geometry providers, the services no longer have
any location set, and in general the plan forward is not to limit
geometer providers by location, so the location check in the lookup is
no longer needed.

To reflect this, we also remove the 'localized' prefix from all config
parameters and related methods/variables.