fix(sz/conmon): show sanction checkboxes in correct steps
This is a fix-up for [1], which implemented the requirement incorrectly:
instead of moving 'endabnahme' to 'Schlussabnahme Gebäude', it did so
for the 'realisierung' step (compare the description in the MR [2] with
the commit message).

This now reverts that change, and now instead moves the 'endabnahme'
sanction step to 'Schlussabnahme Gebäude'.

[1] 26bf9664d5cef723cd50a8a26cba5d2323ff76cf
[2] https://git.adfinis.com/camac-ng/camac-ng/-/merge_requests/10465