feat(security): add censored exception report filter
We use email logging of uncaught exceptions in production environments.
Those emails are currently stuffed with sensitive data that is not
necessary for debugging and should not be sent in such an unsecure way.

This new exception report filter will exclude all of the sensitive data
such as settings and will also censor sensitive request data such as the
authorization token.