fix(be): various performance improvements
Querying the "source" relationship on caluma documents is very
expensive, because each lookup has to go through the visibility layer.
This adds a new property "parent" on the instance API instead, which is
not checked for visibilities (the mere existence of a parent instance
does not have to be a secret). Since the actual instances with the same
eBau number are still loaded through Caluma, this doesn't change which
dossiers are listed.

For a group of 7 instances, this yields a performance improvement from
~8s to ~1s.