feat(alexandria): consistency check for Alexandria files vs S3 bucket
Management command that checks for (1) Alexandria files for which the
corresponding S3 object doesn't exist ('missing') and (2) S3 objects for
which there is no Alexandria file referencing it ('orphans').

To be consumed by a monitoring check to alert/warn operators and/or
developers about potential issues related to Alexandria files.