chore(sz): use config for tax admin notification on instance completion
Originally the code only added the tax administration to the list of
recipients if a checkbox was selected. This was removed with [1] (and
later fixed up with [2]), so the notification is now sent to tax admin
unconditionally (for SZ).

But that means we can now simply declare this in the config, instead of
keeping around canton-specific code.

[1] c63b0dc88b8e2a63c7fb8214fdde2fe84ef97c46
[2] 2afcd90d22ab03ff957ed9651633333eb7699dac