fix(makefile): quote SNAPSHOT var in case it hasn't been set
If SNAPSHOT isn't set, the second line would previously be:

    @[ -n "" ] && docker ... >

(i.e. a redirect without any filename after it), causing a slightly
confusing shell syntax error in addition to the error message printed in
the line above.

Quoting the variable will cause the line to end with ... >"", which
avoids the syntax error.