chore(sz): don't store Ember buildtime configuration in HTML meta field
`storeConfigInMeta` causes the configuration dictionary returned by
config/environment.js::module.exports() to be stored as a meta field in
the app's index.html, to be consumed by the application at runtime.

However, the configuration values are URL-encoded in an HTML file, so
they are not subject to replacement by nginx/replace-env-vars.sh.

Before we can switch to using the prod-env target for our Ember builds,
we therefore need to move the config declaration into a JS file, so the
placeholder values will be correctly replaced (i.e. set
`storeConfigInMeta` to false for SZ as well).