chore(sz/compose): unify container image name declaration
Using :${CI_COMMIT_REF_SLUG:-master} ensures we still get the right tag
for the different environments. The only difference is that we remove
the -${APPLICATION} suffix for prod and staging, but that is redundant
anyway, because the ref slug (branch) already contains the canton name.

Once we move to tag-based releases, we can then also refer to the tag
instead of the ref slug.