feat(be): integrate legal-authority entity
Add new service group and roles for legal-authority.
Permissions are mostly handled through the permissions module with
a new access level for the legal-authority, which can be manually
granted. Some new permission-aware cases are added, to ensure that
certain actions are possible for the new roles before fully migrating
to the permissions module.