chore(sz/ci): versioned builds
The new build step replaces the previous two environment-specific build
steps, and triggers on a version tag rather than a Git branch.

This means we can replace CI_COMMIT_REF_SLUG with CI_COMMIT_TAG for
tagging the container images, and we now adopt the usage of bump-version
for building the ember container image.

We also no longer trigger two pipelines for each environment, but just
one for the release. However, this pipeline will now contain *two* jobs
for the deploy job. When deploying, users are now asked to pay attention
to which environment they wish to deploy a release.