fix(be): fix watermark for exam documents
At the moment the logic for the draft watermarks in pdfs of the exam
documents is not correct.

This fix is intended to ensure, that draft watermarks are only in exam
documents after conversion to PDF if the document is not valid. If
an exam document is valid the draft watermarks should not be displayed
in the generated PDF.