fix(sz): remove Wilken product number configuration
We cannot completely disable the billing config, as it would otherwise
cause the sz_billing_settings fixture to be removed, which in turn will
cause all the tests to fail.

This is a bit of a middle ground: Product numbers are configured for no
service, so it shouldn't complain about missing product numbers in a
given bill, yet the tests can still run successfully.