fix(sz): fix duplicate IRs for documents and form
We have lots of duplicate instance resources in the SZ configuration.
This commit fixes the acute problem that the customer reported but this
will need a proper cleanup when the permission module is being
integrated fully.

Currently, the role "Fachstelle" may see documents and form twice as
they get permission once from the "(documents|form)-read" permission and
once from the configured instance resource role ACLs.

To fix this, we made sure that there is only one instance resource which
meant moving the ACLs from the duplicate IRs to the correct one:

- Documents: Move ACLs from IRs `12` and `35` to `1`
- Form: Move ACLs from IRs `11` and `34` to `3`

Related to camac-ng/kanton-schwyz-projektmanagement#29