feat(ech0211): polymorphic document api endpoint
These endpoints can switch between the document module and alexandria
without change in the presented data model. The resulting API data structure
is exactly the same, and therefore shall contain no surprises to the clients
once we switch over.

The implementation for the camac-documents is based on a SQL view, which
turns out to be simpler to implement than a corresponding ORM query,
especially with the resulting filtering and querying. The Alexandria
models are proxy-modelled to help DRF find the correct "type" (resource
name), as is the Camac model for the category (attachment section).

Note that we don't bother with switching the permissions system:
Alexandria fully uses the new permissions system and resulting
visibilities, while the classic one will use the old visibility code.

Additionally to downloading files from both backends (depending on config),
a multi-file download is implemented as well, allowing clients to fetch
multiple files in a ZIP archive.

Drive-By: We also refactor the ech0211 URLs module to not do any decisions
anymore at module level. This caused a lot of issues with the tests becoming
leaky if you forgot to do the correct module reloading dance.